refactor: Filter Configuration을 FilterConfig.java로 모음
