--- backend/src/main/java/endolphin/backend/global/security/SecurityConfig.java
@@ -1,19 +1,0 @@
-package endolphin.backend.global.security;
-
-import org.springframework.context.annotation.Bean;
-import org.springframework.context.annotation.Configuration;
-import org.springframework.boot.web.servlet.FilterRegistrationBean;
-
-@Configuration
-public class SecurityConfig {
-
-    @Bean
-    public FilterRegistrationBean<JwtAuthFilter> jwtAuthFilterRegistration(JwtAuthFilter filter) {
-        FilterRegistrationBean<JwtAuthFilter> registration = new FilterRegistrationBean<>();
-        registration.setFilter(filter);
-        registration.addUrlPatterns("/*");
-        registration.setOrder(1);
-        return registration;
-    }
-}
-
